test: move isSingleLineDclDs tests into blockParser.test.ts
The helper lives in blockParser, so its tests belong beside it rather than
in a folding-specific file. Relocate the isSingleLineDclDs cases into
blockParser.test.ts and delete foldingRange.test.ts. No behaviour change;
same assertions, same count.
